Fix application tests and create helper controller tests
